PROJECT 3 Multi-Cloud AI Deployment Strategy/requirements.txt
fastapi
uvicorn
pydantic
requests
transformers
scikit-learn
pandas
streamlit
boto3
google-cloud-aiplatform
azure-ai-ml
plotly
